+++ src/components/shared/header/navigation/index.js
@@ -1,0 +1,3 @@
+import Navigation from './navigation';
+
+export default Navigation;
