+++ src/components/shared/header/sidebar/index.js
@@ -1,0 +1,3 @@
+import Sidebar from './sidebar';
+
+export default Sidebar;
